requirements.txt
python-dotenv==1.0.0
newspaper3k==0.2.8
lxml_html_clean==0.1.0
mysql-connector-python==8.1.0
chromadb==1.0.0
openai==0.28.0
requests==2.31.0
numpy==1.24.0
pandas==1.5.3 